1. Adujusting all pins apply to mainbroad 2. merge zf_iic and iic module together. Deleted ZF_iic 3. modified motor, signal, encoder a little bit better.
