+++ peripheral/src/iic.c
@@ -296,6 +296,344 @@ void simiic_init(void)
 
 /*Hardware IIC codes write here*/
 
+//-------------------------------------------------------------------------------------------------------------------
+//  @brief      内部使用用户无需关心
+//  @param      NULL            	
+//  @return     void
+//  Sample usage:               无需用户调用，用户请使用h文件中的宏定义
+//-------------------------------------------------------------------------------------------------------------------
+void iic_delay_us(uint16 x)	//33.1776Mhz
+{
+    uint8 i;
+    while(x--)
+    {
+		i = 9;
+		while (--i);
+    }
+}
+
+
+//-------------------------------------------------------------------------------------------------------------------
+//  @brief      内部使用用户无需关心
+//  @param      NULL            	
+//  @return     void
+//  Sample usage:               无需用户调用，用户请使用h文件中的宏定义
+//-------------------------------------------------------------------------------------------------------------------
+uint8 wait(void)
+{
+    uint16 count = 0;
+    uint8 ret = IIC_SEND_OK;
+    while (!(I2CMSST & 0x40))
+    {
+        iic_delay_us(1);
+        if(count++ >= 30)//等待超过30us，则退出等待。
+        {
+            ret = IIC_SEND_FAIL;
+            break;
+        }
+    }
+    I2CMSST &= ~0x40;
+    return ret;
+}
+
+//-------------------------------------------------------------------------------------------------------------------
+//  @brief      内部使用用户无需关心
+//  @param      NULL            	
+//  @return     void
+//  Sample usage:               无需用户调用，用户请使用h文件中的宏定义
+//-------------------------------------------------------------------------------------------------------------------
+uint8 start(void)
+{
+    uint8 ret;
+    I2CMSCR = 0x01;                             //发送start命令
+    ret = wait();
+    return ret;
+}
+
+//-------------------------------------------------------------------------------------------------------------------
+//  @brief      内部使用用户无需关心
+//  @param      NULL            	
+//  @return     void
+//  Sample usage:               无需用户调用，用户请使用h文件中的宏定义
+//-------------------------------------------------------------------------------------------------------------------
+uint8 send_data(char dat)
+{
+    uint8 ret;
+    I2CTXD = dat;                               //写数据到数据缓冲区
+    I2CMSCR = 0x02;                             //发送SEND命令
+    ret = wait();
+    return ret;
+}
+
+//-------------------------------------------------------------------------------------------------------------------
+//  @brief      内部使用用户无需关心
+//  @param      NULL            	
+//  @return     void
+//  Sample usage:               无需用户调用，用户请使用h文件中的宏定义
+//-------------------------------------------------------------------------------------------------------------------
+uint8 recv_ack(void)
+{
+    uint8 ret;
+    I2CMSCR = 0x03;                             //发送读ACK命令
+    ret = wait();
+    return ret;
+}
+
+//-------------------------------------------------------------------------------------------------------------------
+//  @brief      内部使用用户无需关心
+//  @param      NULL            	
+//  @return     void
+//  Sample usage:               无需用户调用，用户请使用h文件中的宏定义
+//-------------------------------------------------------------------------------------------------------------------
+char recv_data(void)							//接收数据
+{
+    I2CMSCR = 0x04;                             //发送RECV命令
+    wait();
+    return I2CRXD;
+}
+
+//-------------------------------------------------------------------------------------------------------------------
+//  @brief      内部使用用户无需关心
+//  @param      NULL            	
+//  @return     void
+//  Sample usage:               无需用户调用，用户请使用h文件中的宏定义
+//-------------------------------------------------------------------------------------------------------------------
+uint8 send_ack(void)
+{
+	uint8 ret;
+    I2CMSST = 0x00;                             //设置ACK信号
+    I2CMSCR = 0x05;                             //发送ACK命令
+    ret = wait();
+    return ret;
+}
+
+//-------------------------------------------------------------------------------------------------------------------
+//  @brief      内部使用用户无需关心
+//  @param      NULL            	
+//  @return     void
+//  Sample usage:               无需用户调用，用户请使用h文件中的宏定义
+//-------------------------------------------------------------------------------------------------------------------
+uint8 send_nak(void)
+{
+	uint8 ret;
+    I2CMSST = 0x01;                             //设置NAK信号
+    I2CMSCR = 0x05;                             //发送ACK命令
+    ret = wait();
+    return ret;
+}
+
+//-------------------------------------------------------------------------------------------------------------------
+//  @brief      内部使用用户无需关心
+//  @param      NULL            	
+//  @return     void
+//  Sample usage:               无需用户调用，用户请使用h文件中的宏定义
+//-------------------------------------------------------------------------------------------------------------------
+uint8 stop(void)
+{
+    uint8 ret;
+    I2CMSCR = 0x06;                             //发送stop命令
+    ret = wait();
+    return ret;
+}
+
+//-------------------------------------------------------------------------------------------------------------------
+//  @brief      写入一个字节数据到I2C设备指定寄存器地址
+//  @param      iic_n       IIC模块(IIC_1,IIC_2,IIC_3,IIC_0)
+//  @param      slaveid     从机地址(7位地址)
+//  @param      reg         从机寄存器地址
+//  @param      dat         需要发送的数据
+//  @return                 返回的状态值 0：成功  1：失败
+//  @since      v2.0
+//  Sample usage:       	iic_write_reg(0x2D, 0x50,2);     //写入数据2到0x50地址，从机地址为0x2D
+//-------------------------------------------------------------------------------------------------------------------
+uint8 hardiic_write_reg(uint8 dev_add, uint8 reg, uint8 dat)
+{
+    if(start() != IIC_SEND_OK)
+        return IIC_SEND_FAIL;
+    if(send_data((dev_add<<1) | 0x00) != IIC_SEND_OK)
+        return IIC_SEND_FAIL;
+    if(recv_ack() != IIC_SEND_OK)
+        return IIC_SEND_FAIL;
+    if(send_data(reg) != IIC_SEND_OK)
+        return IIC_SEND_FAIL;
+    if(recv_ack() != IIC_SEND_OK)
+        return IIC_SEND_FAIL;
+    if(send_data(dat) != IIC_SEND_OK)
+        return IIC_SEND_FAIL;
+    if(recv_ack() != IIC_SEND_OK)
+        return IIC_SEND_FAIL;
+    if(stop() != IIC_SEND_OK)
+        return IIC_SEND_FAIL;
+
+
+    return IIC_SEND_OK;
+}
+
+//-------------------------------------------------------------------------------------------------------------------
+//  @brief      读取I2C设备指定地址寄存器的数据
+//  @param      iic_n        I2C通道号及引脚
+//  @param      dev_add     从机地址(7位地址)
+//  @param      reg         从机寄存器地址
+//  @param      dat         数据地址
+//  @return                 读取的寄存器值
+//  @since      v1.0
+//  Sample usage:       	uint8 value = iic_read_reg(i2c0, 0x2D, 0x50);//读取0x50地址的数据，从机地址为0x2D
+//-------------------------------------------------------------------------------------------------------------------
+uint8 hardiic_read_reg(uint8 dev_add, uint8 reg, uint8 *dat)
+{
+	if(start() != IIC_SEND_OK)
+        return IIC_SEND_FAIL;
+	
+    if(send_data((dev_add<<1) | 0x00) != IIC_SEND_OK)
+        return IIC_SEND_FAIL;
+    if(recv_ack() != IIC_SEND_OK)
+        return IIC_SEND_FAIL;
+	
+    if(send_data(reg) != IIC_SEND_OK)
+        return IIC_SEND_FAIL;
+    if(recv_ack() != IIC_SEND_OK)
+        return IIC_SEND_FAIL;
+	
+	
+   if(start() != IIC_SEND_OK)
+        return IIC_SEND_FAIL;
+   
+    if(send_data((dev_add<<1) | 0x01) != IIC_SEND_OK)
+        return IIC_SEND_FAIL;
+	
+    if(recv_ack() != IIC_SEND_OK)
+        return IIC_SEND_FAIL;
+	
+
+    *dat = recv_data(); //读取数据
+
+	
+    if(send_nak() != IIC_SEND_OK)
+        return IIC_SEND_FAIL;
+	
+    if(stop() != IIC_SEND_OK)
+        return IIC_SEND_FAIL;
+	
+    return IIC_SEND_OK;
+}
+
+//-------------------------------------------------------------------------------------------------------------------
+//  @brief      读取I2C设备指定地址寄存器的数据
+//  @param      iic_n       I2C通道号及引脚
+//  @param      dev_add     从机地址(7位地址)
+//  @param      reg         从机寄存器地址
+//  @param      dat         读取的数据存储的地址
+//  @param      num         读取字节数
+//  @return     void
+//  @since      v1.0
+//  Sample usage:       	uint8 value = i2c_read_reg(i2c0, 0x2D, 0x50, 10, buf);//读取0x50地址的数据，从机地址为0x2D开始的10个字节
+//-------------------------------------------------------------------------------------------------------------------
+uint8 hardiic_read_reg_bytes(uint8 dev_add, uint8 reg
+						, uint8 *dat, uint8 num)
+{
+
+	if(start() != IIC_SEND_OK)
+        return IIC_SEND_FAIL;
+	
+    if(send_data((dev_add<<1) | 0x00) != IIC_SEND_OK)
+        return IIC_SEND_FAIL;
+    if(recv_ack() != IIC_SEND_OK)
+        return IIC_SEND_FAIL;
+	
+    if(send_data(reg) != IIC_SEND_OK)
+        return IIC_SEND_FAIL;
+    if(recv_ack() != IIC_SEND_OK)
+        return IIC_SEND_FAIL;
+
+	if(start() != IIC_SEND_OK)
+        return IIC_SEND_FAIL;
+		
+	if(send_data((dev_add<<1) | 0x01) != IIC_SEND_OK)
+		return IIC_SEND_FAIL;
+	if(recv_ack() != IIC_SEND_OK)
+		return IIC_SEND_FAIL;
+
+    while(--num)
+    {
+        *dat = recv_data(); //读取数据
+		if(send_ack() != IIC_SEND_OK)
+		{
+			return IIC_SEND_FAIL;
+		}
+        dat++;
+    }
+	
+	*dat = recv_data();
+	
+	if(send_nak() != IIC_SEND_OK)
+		return IIC_SEND_FAIL;
+	
+	if(stop() != IIC_SEND_OK)
+		return IIC_SEND_FAIL;
+	
+	return IIC_SEND_OK;
+}
+
+//-------------------------------------------------------------------------------------------------------------------
+//  @brief      交换I2C设备指定地址寄存器的数据
+//  @param      dev_add     	 设备地址(7位地址)
+//  @param      write_add        写入设备地址
+//  @param      write_num        写入数据的个数
+//  @param      read_add         读取字节地址
+//  @param      read_num         读取字节数
+//  @return     void
+//  @since      v1.0
+//-------------------------------------------------------------------------------------------------------------------
+uint8 hardiic_transfer(const uint8 dev_add, const uint8 *write_add, uint8 write_num, uint8 *read_add, uint8 read_num)
+{
+	if(start() != IIC_SEND_OK)
+		return IIC_SEND_FAIL;
+	
+	if(send_data((dev_add<<1) | 0x00) != IIC_SEND_OK)
+        return IIC_SEND_FAIL;
+	
+	if(recv_ack() != IIC_SEND_OK)
+			return IIC_SEND_FAIL;
+	
+	while(write_num --)
+    {
+		if(send_data(*write_add++) != IIC_SEND_OK)
+			return IIC_SEND_FAIL;
+		if(recv_ack() != IIC_SEND_OK)
+			return IIC_SEND_FAIL;
+    }
+	
+	if(start() != IIC_SEND_OK)
+		return IIC_SEND_FAIL;
+	
+	if(send_data((dev_add<<1) | 0x01) != IIC_SEND_OK)
+        return IIC_SEND_FAIL;
+	
+	if(recv_ack() != IIC_SEND_OK)
+		return IIC_SEND_FAIL;
+	
+    while(--read_num)
+    {
+        *read_add = recv_data(); //读取数据
+		if(send_ack() != IIC_SEND_OK)
+		{
+			return IIC_SEND_FAIL;
+		}
+        read_add++;
+    }
+	
+	*read_add = recv_data();
+	
+	if(send_nak() != IIC_SEND_OK)
+		return IIC_SEND_FAIL;
+	
+	if(stop() != IIC_SEND_OK)
+		return IIC_SEND_FAIL;
+	
+	return IIC_SEND_OK;
+}
+
+
 //========================================================================
 // 函数: void I2C_Init(I2C_InitTypeDef *I2Cx)
 // 描述: I2C初始化程序.
@@ -314,7 +652,6 @@ void hardiic_init()
 	
 	IIC_PIN.Pin = IIC_SDA_PIN;
 	GPIO_Inilize(IIC_GPIO,&IIC_PIN);	
-	
 	//开漏输出设置上拉电阻
 	switch (IIC_GPIO){
 		case GPIO_P1 :P1_PULL_UP_ENABLE(IIC_SCL_PIN | IIC_SDA_PIN);
@@ -326,7 +663,28 @@ void hardiic_init()
 		default : 
 			break;
 	}
-	Zf_IIC_init(IIC_2, IIC2_SCL_P25,IIC2_SDA_P24,63);
+	
+	// 配置IIC寄存器
+	P_SW2 &= ~(0x03<<4);
+    P_SW2 |= 1<<7;	//将EAXFR寄存器置1，这样才能使用特殊功能寄存器为扩展SFR，访问逻辑地址位于 XDATA 区域
+    switch(IIC_GPIO)
+    {
+    case GPIO_P1:
+        P_SW2 |= (0x00<<4);	//SCL:P1.5	SDA:P1.4
+        break;
+    case GPIO_P2:
+        P_SW2 |= (0x01<<4);	//SCL:P2.5	SDA:P2.4
+        break;
+    case GPIO_P3:
+        P_SW2 |= (0x02<<4);	//SCL:P7.7	SDA:P7.6
+        break;
+	default : 
+		break;
+    }
+    I2CCFG |= 1<<6;		//主机模式
+    I2CCFG |= 1<<7;		//使能IIC
+    I2CCFG |= WAIT_TIME;//速度设置为等待wait_time*2+1个时钟
+    I2CMSST = 0x00;		//主机状态寄存器
 }
 
 #endif
@@ -349,29 +707,10 @@ void iic_write_reg(uint8 dev_add, uint8 reg, uint8 dat)
 #if USE_SOFT_IIC
 	simiic_write_reg(dev_add, reg, dat);				/* 设备地址(低七位地址)，存放目标寄存器地址的数组，寄存器数据字节数，写入的数据 */
 #else	//use hardware IIC
-	Zf_IIC_write_reg(dev_add, reg, dat);
+	hardiic_write_reg(dev_add, reg, dat);
 #endif	//USE_SOFT_IIC
 }
 
-
-
-//-------------------------------------------------------------------------------------------------------------------
-//  @brief      IIC写多字节数据
-//  @param      dev_add			设备地址(低七位地址)
-//  @param      reg_add			存放目标寄存器地址的数组
-//	@param		reg_num			寄存器数据字节数
-//  @param      dat_add			要写入的数据保存的地址指针
-//  @param      dat_num			读取字节数量
-//  Sample usage:				
-//-------------------------------------------------------------------------------------------------------------------
-void iic_write_regs(uint8 dev_add, uint8 reg, uint8 *dat_add, uint8 dat_num)
-{
-#if USE_SOFT_IIC
-	simiic_write_regs(dev_add, reg, dat_add, dat_num);
-#else	//use hardware IIC
-	//I2C_WriteNbyte(dev_add, reg_add, reg_num, dat_add, dat_num);
-#endif	//USE_SOFT_IIC
-}
 
 //-------------------------------------------------------------------------------------------------------------------
 //  @brief      IIC读写传输数据
@@ -387,7 +726,7 @@ void iic_transfer(const uint8 dev_add, const uint8 *write_add, uint8 write_num, 
 #if USE_SOFT_IIC
 	simiic_transfer(dev_add, write_add, write_num, read_add, read_num);
 #else	//use hardware IIC
-	Zf_IIC_transfer(dev_add, write_add, write_num, read_add, read_num);
+	hardiic_transfer(dev_add, write_add, write_num, read_add, read_num);
 #endif	//USE_SOFT_IIC
 }
 
@@ -407,7 +746,7 @@ uint8 iic_read_reg(uint8 dev_add, uint8 reg)
 #if USE_SOFT_IIC
 	recive = simiic_read_reg(dev_add, reg);
 #else	//use hardware IIC
-	Zf_IIC_read_reg(dev_add, reg, &recive);
+	hardiic_read_reg(dev_add, reg, &recive);
 #endif	//USE_SOFT_IIC
 	return recive;
 }
@@ -427,7 +766,7 @@ void iic_read_regs(uint8 dev_add, uint8 reg, uint8 *dat_add, uint8 dat_num)
 #if USE_SOFT_IIC
 	simiic_read_regs(dev_add, reg, dat_add, dat_num);
 #else	//use hardware IIC
-	Zf_IIC_read_reg_bytes(dev_add, reg, dat_add, dat_num);
+	hardiic_read_reg_bytes(dev_add, reg, dat_add, dat_num);
 #endif	//USE_SOFT_IIC
 }
