+++ peripheral/src/screen.c
@@ -19,7 +19,7 @@ void HardWare_SPI_Config(void)
 	SPI_Init(&SPI_InitStructure);
 	NVIC_SPI_Init(DISABLE,Priority_3);		//中断配置，是否使用中断
 	
-	SPI_SW(SPI_P22_P23_P24_P25);	//SPI_P54_P13_P14_P15,SPI_P22_P23_P24_P25,SPI_P54_P40_P41_P43,SPI_P35_P34_P33_P32
+	SPI_SW(SPI_P54_P40_P41_P43);	//SPI_P54_P13_P14_P15,SPI_P22_P23_P24_P25,SPI_P54_P40_P41_P43,SPI_P35_P34_P33_P32
     SPI_SS_2 = 1;                   //SS置高电平，取消从机的使能
 }
 
@@ -39,9 +39,13 @@ void  SPI_Simulate_WriteData(u8 Data)
 	}
 }
 
-void GPIO_SPI_Init(uchar GPIO_Pin)
+void GPIO_SPI_Init()
 {
-    P2_MODE_IO_PU(GPIO_Pin);		//P2 设置为准双向口
+	//设置为准双向口
+	P0_MODE_IO_PU(GPIO_Pin_5 | GPIO_Pin_7);
+	P1_MODE_IO_PU(GPIO_Pin_3);
+    P2_MODE_IO_PU(GPIO_Pin_1);		
+	P4_MODE_IO_PU(GPIO_Pin_0 | GPIO_Pin_3);
 }
 
 //向液晶屏写一个8位指令
@@ -235,7 +239,7 @@ void Screen_SetLocation(u16 x,u16 y)
 void Screen_Init(void)
 {
     HardWare_SPI_Config();
-    GPIO_SPI_Init(GPIO_Pin_LOW|GPIO_Pin_5|GPIO_Pin_6);
+    GPIO_SPI_Init();
     Lcd_Init();
 }
 
@@ -577,7 +581,7 @@ uint8 xdata TFT_ASCII[] =
 0x00,0x00,0x10,0x10,0x00,0x00,0x30,0x10,0x10,0x10,0x10,0x10,0x10,0x38,0x00,0x00, //"i"
 0x00,0x00,0x10,0x10,0x00,0x00,0x30,0x10,0x10,0x10,0x10,0x10,0x10,0x90,0x60,0x00, //"j"
 0x00,0x00,0xC0,0x40,0x40,0x40,0x5C,0x48,0x50,0x60,0x50,0x48,0x44,0xEE,0x00,0x00, //"k"
-0x00,0x00,0x30,0x10,0x10,0x10,0x10,0x11,0x10,0x10,0x10,0x10,0x10,0x39,0x00,0x00, //"l"
+0x00,0x00,0x3c,0x10,0x10,0x10,0x10,0x10,0x10,0x10,0x10,0x10,0x10,0x3c,0x00,0x00, //"l"
 0x00,0x00,0x00,0x00,0x00,0x00,0xAC,0xD2,0x92,0x92,0x92,0x92,0x92,0xD6,0x00,0x00, //"m"
 0x00,0x00,0x00,0x00,0x00,0x00,0x58,0xE4,0x44,0x44,0x44,0x44,0x44,0xEE,0x00,0x00, //"n"
 0x00,0x00,0x00,0x00,0x00,0x00,0x38,0x44,0x82,0x82,0x82,0x82,0x44,0x38,0x00,0x00, //"o"
